Remove auth guards and decorators from users controller, move them to separate 'guards' and 'decorators' folders
The following files have been deleted:
- jwt-auth.guard.ts and roles-auth.decor
