+++ src/users/users.controller.ts
@@ -11,9 +11,9 @@ import { CreateUserDto } from './dto/create-user.dto'
 import { UsersService } from './users.service'
 import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger'
 import { User } from './model/users.model'
-import { JwtAuthGuard } from 'src/auth/jwt-auth.guard'
-import { RolesGuard } from 'src/auth/roles.guard'
-import { Roles } from 'src/auth/roles-auth.decorator'
+import { JwtAuthGuard } from 'src/guards/jwt-auth.guard'
+import { RolesGuard } from 'src/guards/roles.guard'
+import { Roles } from 'src/decorators/roles-auth.decorator'
 import { AddRoleDto } from './dto/add-role.dto'
 import { BanUserDto } from './dto/ban-user.dto'
